requirements.txt
django-rest-framework==0.1.0
whitenoise
tzdata
tzlocal